requirements.txt
google-cloud-vision==3.4.5
pathlib==1.0.1
openai==1.3.6
